Refactor in order to enhance coverage
